// src/components/company/interfaces/ICompanyFormState.ts
export interface ICompanyFormState {
    showPassword: boolean;
    showConfirmPassword: boolean;
    companyName: string;
    tradingName: string;
    zipcode: string;
    registerNumber: string;
    name: string;
    email: string;
    login: string;
    password: string;
    confirmPassword: string;
}